Formatted the values on the index page. Dollars now have a dollar sign and percents now have percent sign.

+++ src/main/java/com/inprogress/portfolio/domain/Stock.java
@@ -23,7 +23,7 @@ public class Stock {
     private Long id;
 
     private String stockSymbol;
-    private Integer numberOfShares;
+    private Double numberOfShares;
     private BigDecimal stockPurchasePrice;
     
     // Display only the stock symbol to make debugging easier
